requirements.txt
transformers>=4.30.0
datasets>=2.14.0
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
matplotlib>=3.7.0
seaborn>=0.12.0
accelerate>=0.20.0
tqdm>=4.65.0
nltk>=3.8.0
spacy>=3.5.0
scispacy>=0.5.3
openai>=1.0.0